Final version (before adding visuals)
A game of Rock, Paper, Scissors played in the console. I've added a round counter along with a score counter so that when the game is over a message is logged the details who won along with the final score. I also added a reset function after noticing that the score remained the same after the game was over.